Don't parse command line args in Crucible lib
Moved the parsing of command line args to the main function of
users of the crucible library.  It is now the responsibilty of
main to take care of any command line args and then make sure they
fit into the CrucibleOpts struct.  This allows each binary to have
unique args.
