+++ upstairs/src/lib.rs
@@ -12,7 +12,6 @@ use crucible_protocol::*;
 use anyhow::{anyhow, bail, Result};
 use futures::{SinkExt, StreamExt};
 use ringbuffer::{AllocRingBuffer, RingBufferExt, RingBufferWrite};
-use structopt::StructOpt;
 use usdt::register_probes;
 
 use tokio::net::tcp::WriteHalf;
@@ -24,65 +23,9 @@ use tokio_util::codec::{FramedRead, FramedWrite};
 // Include the Rust implementation generated by the build script.
 include!(concat!(env!("OUT_DIR"), "/crutrace.rs"));
 
-#[derive(Debug, StructOpt)]
-#[structopt(about = "volume-side storage component")]
-pub struct Opt {
-    #[structopt(short, long, default_value = "127.0.0.1:9000")]
-    target: Vec<SocketAddrV4>,
-}
-
-pub fn opts() -> Result<Opt> {
-    let opt: Opt = Opt::from_args();
-    println!("raw options: {:?}", opt);
-
-    if opt.target.is_empty() {
-        bail!("must specify at least one --target");
-    }
-
-    Ok(opt)
-}
-
-impl Opt {
-    /*
-     * Use:
-     *
-     *     let opt = Opt::from_string("-- -t 192.168.1.1:3801 -t 192.168.1.2:3801".to_string()).unwrap();
-     *
-     */
-    pub fn from_string(args: String) -> Result<Opt> {
-        let opt: Opt = Opt::from_iter(args.split(' '));
-
-        if opt.target.is_empty() {
-            bail!("must specify at least one --target");
-        }
-
-        Ok(opt)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::Opt;
-    use std::net::{Ipv4Addr, SocketAddrV4};
-
-    #[test]
-    fn test_opt_from_string() {
-        let opt = Opt::from_string(
-            "-- -t 192.168.1.1:3801 -t 192.168.1.2:3801".to_string(),
-        )
-        .unwrap();
-        assert_eq!(opt.target.is_empty(), false);
-        assert_eq!(opt.target.len(), 2);
-
-        assert_eq!(
-            opt.target[0],
-            SocketAddrV4::new(Ipv4Addr::new(192, 168, 1, 1), 3801)
-        );
-        assert_eq!(
-            opt.target[1],
-            SocketAddrV4::new(Ipv4Addr::new(192, 168, 1, 2), 3801)
-        );
-    }
+#[derive(Debug)]
+pub struct CrucibleOpts {
+    pub target: Vec<SocketAddrV4>,
 }
 
 pub fn deadline_secs(secs: u64) -> Instant {
@@ -834,7 +777,7 @@ pub struct Upstairs {
 }
 
 impl Upstairs {
-    pub fn new(opt: &Opt, guest: Arc<Guest>) -> Arc<Upstairs> {
+    pub fn new(opt: &CrucibleOpts, guest: Arc<Guest>) -> Arc<Upstairs> {
         Arc::new(Upstairs {
             guest,
             ds_work: Mutex::new(Work {
@@ -1834,7 +1777,7 @@ async fn up_listen(up: &Arc<Upstairs>, dst: Vec<Target>) {
  * XXX At the moment, this function is only half complete, and will
  * probably need a re-write.
  */
-pub async fn up_main(opt: Opt, guest: Arc<Guest>) -> Result<()> {
+pub async fn up_main(opt: CrucibleOpts, guest: Arc<Guest>) -> Result<()> {
     match register_probes() {
         Ok(()) => {
             println!("DTrace probes registered ok");
